Wire in the basic authentication and token service, with a simple in-memory UUID token. Support login token and detecting when the token is no longer valid in the underlying webapp - I'm not convinced the window.location.reload() is the right approach, but we can revisit.

+++ src/main/java/org/attendantsoffice/eventmanager/user/UserController.java
@@ -23,7 +23,7 @@ public class UserController {
     }
 
     @GetMapping(path = "/api/users/{userId}")
-    public UserEntity findUser(@PathVariable Long userId) {
+    public UserEntity findUser(@PathVariable Integer userId) {
         return userRepository.findById(userId).orElseThrow(() -> new IllegalArgumentException("No User#" + userId
                 + " found"));
     }
