+++ src/main/java/org/attendantsoffice/eventmanager/user/UserRepository.java
@@ -9,7 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Repository
 @Transactional
-public interface UserRepository extends CrudRepository<UserEntity, Long> {
+public interface UserRepository extends CrudRepository<UserEntity, Integer> {
 
     @Transactional(readOnly = true)
     Optional<UserEntity> findByEmail(String email);
